Move some mutations.test to the client side
This has to be done on the client side because on the server side we do
not have access to a full client with a cache. Only the tests using
the `multiQueryUpdater` have been moved.

Linked to #2415.
