+++ packages/vulcan-core/test/client/index.js
@@ -1,1 +1,3 @@
 import '../index';
+
+import './mutations2.test';
